chores: move internal Now() to new helper package
