+++ internal/helper/time_testing.go
@@ -1,10 +1,10 @@
 //go:build testing
 // +build testing
 
-package tasks
+package helper
 
 import "time"
 
-func Time() time.Time {
+func Now() time.Time {
 	return time.Date(2000, time.January, 1, 0, 0, 0, 0, time.UTC)
 }
